Moved all backend api files to api folder within react project for convienience
