--- frontend/src/components/pages/About.jsx
@@ -1,9 +1,0 @@
-import './_pages.scss';
-
-const About = () => {
-  return (
-    <div className='about'></div>
-  );
-}
-
-export default About;
