--- frontend/src/components/pages/AuthPage.jsx
@@ -1,9 +1,0 @@
-import './_pages.scss';
-
-const AuthPage = () => {
-  return (
-    <div className='auth-page'></div>
-  );
-}
-
-export default AuthPage;
