--- frontend/src/components/pages/Cart.jsx
@@ -1,9 +1,0 @@
-import './_pages.scss';
-
-const Cart = () => {
-  return (
-    <div className='cart'></div>
-  );
-}
-
-export default Cart;
