--- frontend/src/components/pages/Home.jsx
@@ -1,9 +1,0 @@
-import './_pages.scss';
-
-const Home = () => {
-  return (
-    <div className='home'>Hello</div>
-  );
-}
-
-export default Home;
