--- frontend/src/components/pages/Products.jsx
@@ -1,9 +1,0 @@
-import './_pages.scss';
-
-const Products = () => {
-  return (
-    <div className='products'></div>
-  );
-}
-
-export default Products;
