--- frontend/src/components/pages/_pages.scss
@@ -1,12 +1,0 @@
-// Not Found page
-.not-found {
-    display:flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    min-height: 80vh;
-    width:100%;
-    .h1,.h3 {
-        color: black;
-    }
-}
