--- frontend/src/index.scss
@@ -1,9 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Nunito:wght@200;400;700&display=swap');
-
-* {
-  text-decoration: none !important;
-  box-sizing: border-box;
-  padding: 0;
-  margin:0;
-  font-family: 'Nunito', 'Roboto', sans-serif;
-}
