--- src/api/products.js
@@ -1,49 +1,0 @@
-const products = [
-    {
-        id:1,
-        item:"Signed Jersey",
-        name:"Ichiro Suzuki",
-        category: ["sports", "baseball"],
-        authenticated:"yes",
-        price: 199.99,
-        image: ""
-    },
-    {
-        id:2,
-        item:"Signed Football",
-        name:"Steve Largent",
-        category: ["sports", "football"],
-        authenticated:"yes",
-        price: 99.99,
-        image: ""
-    },
-    {
-        id:3,
-        item:"Signed Helmet",
-        name:"Deion \"Primetime\" Sanders",
-        category: ["sports", "football"],
-        authenticated:"yes",
-        price: 199.99,
-        image: ""
-    },
-    {
-        id:4,
-        item:"Quill",
-        name:"Ben Franklin",
-        category: ["history", "politics"],
-        authenticated:"yes",
-        price: 599.99,
-        image: ""
-    },
-    {
-        id:5,
-        item:"Used \"Snot-rag\"",
-        name:"Justin Bieber",
-        category: ["celebrity", "personal"],
-        authenticated:"no",
-        price: 1299.99,
-        image: ""
-    }
-]
-
-module.exports = products;
